Refactor to use platform abstraction for memory operations and type definitions
- Replaced standard library calls (`memset`, `memcpy`, etc.) with platform-specific wrappers (`sf_memset`, `sf_memcpy`, etc.).
- Consolidated type definitions and removed system headers in favor of `starforth_platform.h`.
- Simplified logging and IO setup by eliminating dependency on standard headers like `string.h` and `stdarg.h`.
- Streamlined VM memory initialization using platform abstractions.

+++ include/vm.h
@@ -18,14 +18,14 @@
 #ifndef VM_H
 #define VM_H
 
-/* ANSI C99 only - minimal headers */
-#include <stddef.h>
-#include <stdint.h>
+/* Bare metal - no system headers */
+#include "platform/starforth_platform.h"
 
 /* Forward declaration */
 struct VM;
 
-typedef intptr_t cell_t;
+/* Bare metal type definitions */
+typedef signed long cell_t;
 
 
 typedef void (*word_func_t)(struct VM *vm);
